Compiling error fix in TEncSlice.cpp

+++ branches/HM-14.0-KTA-1.0-dev/source/Lib/TLibEncoder/TEncTop.cpp
@@ -684,7 +684,7 @@ Void TEncTop::xInitSPS()
 
   m_cSPS.setMaxTLayers( m_maxTempLayer );
   m_cSPS.setTemporalIdNestingFlag( ( m_maxTempLayer == 1 ) ? true : false );
-  for ( i = 0; i < min(m_cSPS.getMaxTLayers(),(UInt) MAX_TLAYER); i++ )
+  for ( i = 0; i < min((Int)m_cSPS.getMaxTLayers(),(Int) MAX_TLAYER); i++ )
   {
     m_cSPS.setMaxDecPicBuffering(m_maxDecPicBuffering[i], i);
     m_cSPS.setNumReorderPics(m_numReorderPics[i], i);
